fix/#4-mov-card-buttons-to-card-itself кнопки редактирования карточки вынесены на верхний уровень
